add userService interface and implementation along with some role methods
