+++ src/main/java/com/myprojects/demo/repositories/RoleRepository.java
@@ -4,4 +4,5 @@ import com.myprojects.demo.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findByName(String name);
 }
