+++ src/main/java/com/myprojects/demo/services/UserService.java
@@ -1,40 +1,13 @@
 package com.myprojects.demo.services;
 
 import com.myprojects.demo.entities.MovieUser;
-import com.myprojects.demo.exceptions.UsernameException;
-import com.myprojects.demo.repositories.UserRepository;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Service;
+import com.myprojects.demo.entities.Role;
 
-import javax.transaction.Transactional;
 import java.util.List;
-import java.util.Optional;
 
-@Service
-public class UserService {
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
-    private final UserRepository userRepository;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    public List<MovieUser> getAllUsers() {
-        return userRepository.findAll();
-    }
-
-    @Transactional
-    public MovieUser addUser(String username, String password) {
-        MovieUser movieUser = new MovieUser();
-        movieUser.setUsername(username);
-        movieUser.setPassword(password);
-        Optional<MovieUser> existingUser = userRepository.findByUsername(username);
-        if (existingUser.isPresent()) {
-            throw new UsernameException("Username: " + username + " already exists.");
-        }
-        movieUser = userRepository.save(movieUser);
-        log.info("New user: {} added.", username);
-        return movieUser;
-    }
+public interface UserService {
+    List<MovieUser> getAllUsers();
+    MovieUser addUser(String username, String password);
+    Role saveRole(Role role);
+    void addRoleToUser(String username, String roleName);
 }
